+++ backend/src/app.module.ts
@@ -1,8 +1,10 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { TypeOrmModule } from '@nestjs/typeorm';
+
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
+import { GenreModule } from './modules/genre/genre.module';
 
 @Module({
   imports: [
@@ -20,10 +22,13 @@ import { AppService } from './app.service';
         password: config.get<string>('DATABASE_PASSWORD'),
         database: config.get<string>('DATABASE_NAME'),
         entities: [__dirname + '/**/*.entity{.ts,.js}'],
+        migrations: [__dirname + '/../database/migrations/*{.ts,.js}'],
+        migrationsRun: false,
         synchronize: false,
         logging: config.get<string>('NODE_ENV') === 'development',
       }),
     }),
+    GenreModule,
   ],
   controllers: [AppController],
   providers: [AppService],